+++ src/main/java/uk/org/binky/lexinator/Token.java
@@ -1,5 +1,10 @@
 package uk.org.binky.lexinator;
 
+/**
+ * This class describes a token.
+ *
+ * @param <T> The token type used.
+ */
 public final class Token<T extends Enum<T>> {
 	public final String file;
 	public final int line;
@@ -11,7 +16,14 @@ public final class Token<T extends Enum<T>> {
 		this.type = type;
 		this.value = value;
 	}
-	
+
+	/**
+	 * Compares this token to another, and returns true if they are equal.
+	 * TODO: proper compareTo, equals, hash, etc.
+	 *
+	 * @param that Another token
+	 * @return true if both tokens are equal
+	 */
 	public boolean compare(final Token<T> that) {
 		if (this.file != that.file || this.line != that.line || this.type != that.type || !this.value.equals(that.value)) {
 			return false;