+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.0
+- [x] Lexer.name and Lexer.text are now private
+- [x] Lexer.state is no longer protected, but private
+- [x] Lexer.setState is no longer protected, but private
+- [x] Lexer's constructors are not protected
+- [x] Removed deprecated whitespace methods
+
 ## v0.1.4
 
 - [x] Make Token constructor public.